refactor: :recycle: refactor server init files
