+++ docker/docker-compose.ci.yml
@@ -48,7 +48,7 @@ services:
     ports:
       - 4000:80
     volumes:
-      - ../packages/test-utils/src/imports:/app/src/init-db/imports
+      - ../packages/test-utils/src/imports:/app/src/init/db/imports
     environment:
       SERVER_PORT: 80 # ${SERVER_PORT:-4000}
       DB_HOST: ${POSTGRES_HOST:-postgres}
